[SYCL][HIP] Update HIP AMD XFAILs
The get_backend test was missing the HIP plugin case.

The ESIMD tests are not supported for HIP in general, not just on
Nvidia.

The subgroup built-ins were added for AMD in intel/llvm#4208.

The kernel_info.cpp test was fixed in #474 and now passes with HIP.

The kernel_bundle_ignore_sycl_external.cpp only fails with HIP Nvidia,
not on AMD.