[SYCL][CUDA] Enable queue barrier tests
These are passing with CUDA, as far as I can tell this feature was
implemented as part of intel/llvm#3932.

`enqueue_barrier.cpp` had to be tweaked to accept a triple as that is
required when using the CUDA plugin.

Note that this feature is missing from the hip plugin which is why these
are still unsupported for hip.